# Conflicts:
#	main.py

+++ main.py
@@ -1,5 +1,7 @@
 import os
 import discord
+
+from discord.utils import get
 from replies import rick_reply
 from dotenv import load_dotenv
 from discord.ext import commands
@@ -30,11 +32,14 @@ async def on_message(message):
         # await rick.process_commands(message)
 
 
-@rick.event
+@rick.event  #command for deleting a message when a set amount of reactions have been added
 async def on_raw_reaction_add(payload):
     channel = await rick.fetch_channel(payload.channel_id)
     message = await channel.fetch_message(payload.message_id)
-    await message.delete()
+    if payload.emoji.name == "❌":
+        reaction = get(message.reactions, emoji=payload.emoji.name)
+    if reaction and reaction.count == 4:
+        await message.delete()
 
 
 @rick.command()
